+++ modules/elements/elements/hd-image-comparison/templates/template.php
@@ -66,6 +66,10 @@ $slider = $this->el('div', [
         '{slider_onmousemove}',
     ],
 
+    'data-margin-remove' => [
+        '{slider_margin_remove}',
+    ],
+
 ]);
 
 // Box decoration